provide spirv-llvm-translator.bst for 22.08
mesa main introduced a build time (and runtime) requirement towards LLVMSPIRVLib. we satisfy the build requirement with this knowing that the runtime requirement is not satisfied as the dependency is missing in runtime/platform.

The runtime requirement will most likely not be provided for 22.08. The sole user of this runtime requirement are the Vulkan validation layers.

The vendored spirv-llvm-translator.bst is 22.08 latest spirv-llvm-translator.bst + !13261 from freedesktop-sdk (MR providing the LLVMSPIRVLib build time dependency for 23.08).